+++ src/core/message/handler.js
@@ -1,13 +1,13 @@
 // imports
 import * as middlewares from './middlewares'
-import Message from './sender'
+import * as initialize from './initialize'
 import { errorLog } from '../log'
 /**
  * handles a command by a specific target.
  * @param {object} param A cqhttp message object
  */
 export default async function handle(param) {
-    const msg = new Message(param)
+    const msg = initialize[param.post_type](param)
     if (middlewares[param.post_type])
         for (let middleware of middlewares[param.post_type]) {
             try { await middleware(msg) }