+++ src/core/message/sender/event.js
@@ -1,0 +1,5 @@
+import Message from './message'
+export default class EventMessage extends Message {
+    constructor(param) { super(param) }
+    // TODO: write something to process events.
+}