+++ src/core/message/sender/index.js
@@ -1,0 +1,4 @@
+export Message from './message'
+export ContentMessage from './content'
+export RequestMessage from './request'
+export EventMessage from './event'