+++ src/modules/utility/commands/test.js
@@ -6,7 +6,7 @@ export default {
     options: [],
     /**
      * Returns the text directly to the user
-     * @param {Message} msg The universal msg object
+     * @param {ContentMessage} msg The universal msg object
      * @param {{ txt: string[] }} txt The texts user sends
      */
     async action(msg, { txt }) { msg.send(txt.filter(i => new RegExp(...injectionChecker).test(i) === false).join(' ')) }